# Codegen: don't use init check for consts that are declared before read