--- a/src/compiler/crystal/codegen/const.cr
+++ b/src/compiler/crystal/codegen/const.cr
@@ -96,9 +96,43 @@ class Crystal::CodeGenVisitor
     end
   end
 
-  def initialize_const(const)
-    # Maybe the constant was simple and doesn't need a real initialization
+  def initialize_no_init_flag_const(const)
+    global = declare_const(const)
 
+    with_cloned_context do
+      # "self" in a constant is the constant's namespace
+      context.type = const.namespace
+
+      # Start with fresh variables
+      context.vars = LLVMVars.new
+
+      alloca_vars const.vars
+      request_value do
+        accept const.value
+      end
+    end
+
+    const_type = const.value.type
+    if const_type.passed_by_value?
+      @last = load @last
+    end
+
+    store @last, global
+
+    global.initializer = @last.type.null
+
+    global
+  end
+
+  def initialize_const(const)
+    # If the constant wasn't read yet, we can initialize it right now and
+    # avoid checking an "initialized" flag every time we read it.
+    unless const.read?
+      const.no_init_flag = true
+      return initialize_no_init_flag_const(const)
+    end
+
+    # Maybe the constant was simple and doesn't need a real initialization
     global, initialized_flag = declare_const_and_initialized_flag(const)
     return global if const.initializer
 
@@ -173,7 +207,9 @@ class Crystal::CodeGenVisitor
   end
 
   def read_const_pointer(const)
-    if const == @program.argc || const == @program.argv || const.initializer
+    const.read = true
+
+    if const == @program.argc || const == @program.argv || const.initializer || const.no_init_flag?
       global_name = const.llvm_name
       global = declare_const(const)